requirements.txt
# core ML - torch handled by base docker image, skip pinning
causal-conv1d>=1.6.1
mamba-ssm==2.3.1
einops==0.8.2
transformers==4.45.0
tokenizers==0.20.3
safetensors==0.7.0
datasets==4.8.4
evaluate==0.4.6
huggingface_hub==0.36.2
hf-xet==1.4.3
accelerate

# attention / custom installs
-e git+https://github.com/fla-org/flash-linear-attention.git@e5c18bf633bb093a38ad915c2cfe3979425ab27c#egg=flash_linear_attention
-e git+https://github.com/YaxitaAmin/log-linear-attention-lamda-MLP.git@bacae3ddb0afb5a8d9b6cff1a9a871570bf3430a#egg=hattention

# training utilities
lightning==2.6.1
pytorch-lightning==2.6.1
lightning-utilities==0.15.3
torchmetrics==1.9.0
wandb==0.26.0

# data / utils
numpy==2.4.4
pandas==3.0.2
pyarrow==23.0.1
dill==0.4.1
multiprocess==0.70.19
xxhash==3.6.0
fsspec==2026.2.0

# config
hydra-core==1.3.2
omegaconf==2.3.0
PyYAML==6.0.3

# misc
tqdm==4.67.3
requests==2.33.1
packaging==26.1
filelock==3.28.0
sympy==1.13.1
networkx==3.6.1
ninja==1.13.0
matplotlib==3.10.8
rich==15.0.0
click==8.3.2
GitPython==3.1.46
psutil==7.2.2
regex==2026.4.4
jaxtyping==0.3.9
lark==1.3.1
